vendex-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
asyncpg==0.30.0
httpx==0.27.0
yt-dlp==2025.3.31
opencv-python-headless
numpy==2.1.3
Pillow
python-multipart==0.0.9
python-dotenv==1.0.1
aiofiles==23.2.1
asyncio-throttle==1.0.2
tenacity==8.3.0
pydantic==2.10.3
pydantic-settings==2.6.1
openpyxl==3.1.5
playwright-stealth==2.0.3
forex-python==1.9.2
